backend/pyproject.toml
[project]
name = "trickyclip-backend"
version = "0.1.0"
description = "Backend for TrickyClip"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.9",
    "redis[hiredis]>=5.0.1",
    "redis[asyncio]>=5.0.1",
    "rq>=1.16.0",
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.1",
    "google-api-python-client>=2.110.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "requests>=2.31.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "alembic>=1.13.0",
    "tflite-runtime>=2.14.0",
    "psutil>=5.9.0",
    "librosa>=0.9.2",
    "soundfile>=0.12.1",
    "ffmpeg-python>=0.2.0",
    "cryptography>=41.0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

